+++ lib/model/Product.g.dart
@@ -110,11 +110,15 @@ Product _$ProductFromJson(Map<String, dynamic> json) => Product(
               (k, e) => MapEntry($enumDecode(_$ImageFieldEnumMap, k), e as int),
             )),
       )
+      ..comparedToCategory = json['compared_to_category'] as String?
       ..packagingTextInLanguages =
           LanguageHelper.fromJsonStringMap(json['packaging_text_in_languages'])
       ..knowledgePanels =
           KnowledgePanels.fromJsonHelper(json['knowledge_panels'] as Map?)
-      ..environmentInfoCard = json['environment_infocard'] as String?;
+      ..environmentInfoCard = json['environment_infocard'] as String?
+      ..embCodes = json['emb_codes'] as String?
+      ..manufacturingPlaces = json['manufacturing_places'] as String?
+      ..origins = json['origins'] as String?;
 
 Map<String, dynamic> _$ProductToJson(Product instance) {
   final val = <String, dynamic>{
@@ -176,6 +180,7 @@ Map<String, dynamic> _$ProductToJson(Product instance) {
   writeNotNull('nutriment_energy_unit', instance.nutrimentEnergyUnit);
   writeNotNull('nutrition_data_per', instance.nutrimentDataPer);
   writeNotNull('nutrition_grade_fr', instance.nutriscore);
+  writeNotNull('compared_to_category', instance.comparedToCategory);
   writeNotNull('categories', instance.categories);
   writeNotNull('categories_tags', instance.categoriesTags);
   writeNotNull('categories_tags_in_languages',
@@ -204,6 +209,9 @@ Map<String, dynamic> _$ProductToJson(Product instance) {
   writeNotNull('knowledge_panels',
       KnowledgePanels.toJsonHelper(instance.knowledgePanels));
   writeNotNull('environment_infocard', instance.environmentInfoCard);
+  writeNotNull('emb_codes', instance.embCodes);
+  writeNotNull('manufacturing_places', instance.manufacturingPlaces);
+  writeNotNull('origins', instance.origins);
   val['no_nutrition_data'] =
       JsonHelper.checkboxToJSON(instance.noNutritionData);
   writeNotNull('nutriments', Nutriments.toJsonHelper(instance.nutriments));