+++ lib/utils/HttpHelper.dart
@@ -3,7 +3,6 @@ import 'dart:convert';
 
 import 'package:http/http.dart' as http;
 import 'package:meta/meta.dart';
-
 import 'package:openfoodfacts/openfoodfacts.dart';
 import 'package:openfoodfacts/utils/UriReader.dart';
 import 'package:path/path.dart';
@@ -44,6 +43,14 @@ class HttpHelper {
     if (OpenFoodAPIConfiguration.uuid != null) {
       map ??= <String, String>{};
       map['app_uuid'] = OpenFoodAPIConfiguration.uuid!;
+    }
+    if (OpenFoodAPIConfiguration.userAgent?.system != null) {
+      map ??= <String, String>{};
+      map['app_platform'] = OpenFoodAPIConfiguration.userAgent?.system ?? '';
+    }
+    if (OpenFoodAPIConfiguration.userAgent?.comment != null) {
+      map ??= <String, String>{};
+      map['comment'] = OpenFoodAPIConfiguration.userAgent?.comment ?? '';
     }
     return map;
   }
@@ -167,11 +174,11 @@ class HttpHelper {
       'Accept': 'application/json',
       'UserAgent':
           OpenFoodAPIConfiguration.userAgent?.toValueString() ?? USER_AGENT,
-      'From': OpenFoodAPIConfiguration.getUser(user)?.toValueString() ?? FROM,
+      'From': OpenFoodAPIConfiguration.getUser(user)?.userId ?? FROM,
     });
 
     if (isTestModeActive) {
-      var token = 'Basic ' + base64Encode(utf8.encode('off:off'));
+      var token = 'Basic ${base64Encode(utf8.encode('off:off'))}';
       headers.addAll({'authorization': token});
     }
     return headers;