+++ test/api_getProduct_test.dart
@@ -4,13 +4,6 @@ import 'package:openfoodfacts/model/AttributeGroup.dart';
 import 'package:openfoodfacts/model/NutrientLevels.dart';
 import 'package:openfoodfacts/model/Nutriments.dart';
 import 'package:openfoodfacts/openfoodfacts.dart';
-import 'package:openfoodfacts/personalized_search/available_attribute_groups.dart';
-import 'package:openfoodfacts/personalized_search/available_preference_importances.dart';
-import 'package:openfoodfacts/personalized_search/available_product_preferences.dart';
-import 'package:openfoodfacts/personalized_search/matched_product.dart';
-import 'package:openfoodfacts/personalized_search/preference_importance.dart';
-import 'package:openfoodfacts/personalized_search/product_preferences_manager.dart';
-import 'package:openfoodfacts/personalized_search/product_preferences_selection.dart';
 import 'package:openfoodfacts/utils/CountryHelper.dart';
 import 'package:openfoodfacts/utils/InvalidBarcodes.dart';
 import 'package:openfoodfacts/utils/OpenFoodAPIConfiguration.dart';
@@ -22,10 +15,10 @@ import 'package:test/test.dart';
 import 'test_constants.dart';
 
 void main() {
-  const int _HTTP_OK = 200;
+  const int HTTP_OK = 200;
 
-  const _BARCODE_UNKNOWN = '11111111111111111111111111';
-  const _BARCODE_DANISH_BUTTER_COOKIES = '5701184005007';
+  const BARCODE_UNKNOWN = '11111111111111111111111111';
+  const BARCODE_DANISH_BUTTER_COOKIES = '5701184005007';
 
   OpenFoodAPIConfiguration.globalQueryType = QueryType.TEST;
 
@@ -203,7 +196,7 @@ void main() {
 
     test('get product Danish Butter Cookies & Chocolate Chip Cookies',
         () async {
-      String barcode = _BARCODE_DANISH_BUTTER_COOKIES;
+      String barcode = BARCODE_DANISH_BUTTER_COOKIES;
       ProductQueryConfiguration configurations = ProductQueryConfiguration(
           barcode,
           language: OpenFoodFactsLanguage.GERMAN,
@@ -360,9 +353,9 @@ void main() {
         user: user,
         queryType: queryType,
       );
-      expect(result.product!.nutriments!.chloride, .015);
+      expect(result.product!.nutriments!.chloride, .0015);
       expect(result.product!.nutriments!.chlorideUnit, Unit.MILLI_G);
-      expect(result.product!.nutriments!.chlorideServing, .15);
+      expect(result.product!.nutriments!.chlorideServing, .015);
 
       result = await OpenFoodAPIClient.getProduct(
         ProductQueryConfiguration(
@@ -532,7 +525,7 @@ void main() {
     });
 
     test('product not available', () async {
-      String barcode = _BARCODE_UNKNOWN;
+      String barcode = BARCODE_UNKNOWN;
       ProductQueryConfiguration configurations = ProductQueryConfiguration(
           barcode,
           language: OpenFoodFactsLanguage.GERMAN,
@@ -901,7 +894,7 @@ void main() {
     test(
         'vegan, vegetarian and palm oil ingredients of Danish Butter Cookies & Chocolate Chip Cookies',
         () async {
-      String barcode = _BARCODE_DANISH_BUTTER_COOKIES;
+      String barcode = BARCODE_DANISH_BUTTER_COOKIES;
       ProductQueryConfiguration configurations = ProductQueryConfiguration(
           barcode,
           language: OpenFoodFactsLanguage.GERMAN,
@@ -916,96 +909,6 @@ void main() {
       expect(vegetableFat.vegan, IngredientSpecialPropertyStatus.POSITIVE);
       expect(vegetableFat.vegetarian, IngredientSpecialPropertyStatus.POSITIVE);
       expect(vegetableFat.fromPalmOil, IngredientSpecialPropertyStatus.MAYBE);
-    });
-
-    test('matched product', () async {
-      final Map<String, String> attributeImportances = {};
-      int refreshCounter = 0;
-      final ProductPreferencesManager manager = ProductPreferencesManager(
-        ProductPreferencesSelection(
-          setImportance: (String attributeId, String importanceIndex) async {
-            attributeImportances[attributeId] = importanceIndex;
-          },
-          getImportance: (String attributeId) =>
-              attributeImportances[attributeId] ??
-              PreferenceImportance.ID_NOT_IMPORTANT,
-          notify: () => refreshCounter++,
-        ),
-      );
-      const OpenFoodFactsLanguage language = OpenFoodFactsLanguage.ENGLISH;
-      final String languageCode = language.code;
-      final String importanceUrl =
-          AvailablePreferenceImportances.getUrl(languageCode);
-      final String attributeGroupUrl =
-          AvailableAttributeGroups.getUrl(languageCode);
-      http.Response response;
-      response = await http.get(Uri.parse(importanceUrl));
-      expect(response.statusCode, _HTTP_OK);
-      final String preferenceImportancesString = response.body;
-      response = await http.get(Uri.parse(attributeGroupUrl));
-      expect(response.statusCode, _HTTP_OK);
-      final String attributeGroupsString = response.body;
-      manager.availableProductPreferences =
-          AvailableProductPreferences.loadFromJSONStrings(
-        preferenceImportancesString: preferenceImportancesString,
-        attributeGroupsString: attributeGroupsString,
-      );
-      expect(refreshCounter, 0);
-
-      const String barcode = '0028400047685';
-      final ProductQueryConfiguration configurations =
-          ProductQueryConfiguration(
-        barcode,
-        language: language,
-        fields: [ProductField.NAME, ProductField.ATTRIBUTE_GROUPS],
-      );
-      final ProductResult result = await OpenFoodAPIClient.getProduct(
-        configurations,
-        user: TestConstants.TEST_USER,
-      );
-      expect(result.status, 1);
-      expect(result.barcode, barcode);
-
-      final String attributeId1 = Attribute.ATTRIBUTE_NUTRISCORE;
-      final String attributeId2 = Attribute.ATTRIBUTE_FOREST_FOOTPRINT;
-      final String importanceId1 = PreferenceImportance.ID_MANDATORY;
-      final String importanceId2 = PreferenceImportance.ID_IMPORTANT;
-      await manager.setImportance(attributeId1, importanceId1);
-      expect(
-          manager.getImportanceIdForAttributeId(attributeId1), importanceId1);
-      expect(refreshCounter, 1);
-      await manager.setImportance(attributeId2, importanceId2);
-      expect(
-          manager.getImportanceIdForAttributeId(attributeId2), importanceId2);
-      expect(refreshCounter, 2);
-      MatchedProduct matchedProduct;
-
-      matchedProduct = MatchedProduct(result.product!, manager);
-      assert(matchedProduct.score > 151);
-      expect(matchedProduct.status, MatchedProductStatus.YES);
-
-      await manager.setImportance(attributeId1, importanceId2);
-      expect(
-          manager.getImportanceIdForAttributeId(attributeId1), importanceId2);
-      expect(refreshCounter, 3);
-      await manager.setImportance(attributeId2, importanceId1);
-      expect(
-          manager.getImportanceIdForAttributeId(attributeId2), importanceId1);
-      expect(refreshCounter, 4);
-
-      matchedProduct = MatchedProduct(result.product!, manager);
-      assert(matchedProduct.score > 37.5);
-      expect(
-          matchedProduct.status,
-          MatchedProductStatus
-              .YES); // because the score for FOREST is not good enough
-
-      await manager.clearImportances(); // no attribute parameters at all
-      expect(refreshCounter, 5);
-
-      matchedProduct = MatchedProduct(result.product!, manager);
-      expect(matchedProduct.score, 0.0);
-      expect(matchedProduct.status, MatchedProductStatus.YES);
     });
 
     test(
@@ -1243,7 +1146,7 @@ void main() {
           equals({
             OpenFoodFactsLanguage.RUSSIAN: ['Россия']
           }));
-    });
+    }, skip: 'Random results');
 
     test('multiple languages and in-languages fields', () async {
       String barcode = '2222222222224';
@@ -1564,7 +1467,7 @@ void main() {
     test('get ecoscore html description', () async {
       final ProductResult productResult = await OpenFoodAPIClient.getProduct(
         ProductQueryConfiguration(
-          _BARCODE_DANISH_BUTTER_COOKIES,
+          BARCODE_DANISH_BUTTER_COOKIES,
           language: OpenFoodFactsLanguage.FRENCH,
           fields: <ProductField>[ProductField.ENVIRONMENT_INFOCARD],
         ),
@@ -1583,7 +1486,7 @@ void main() {
       };
       final ProductResult productResult = await OpenFoodAPIClient.getProduct(
         ProductQueryConfiguration(
-          _BARCODE_DANISH_BUTTER_COOKIES,
+          BARCODE_DANISH_BUTTER_COOKIES,
           language: OpenFoodFactsLanguage.FRENCH,
           fields: <ProductField>[ProductField.KNOWLEDGE_PANELS],
         ),
@@ -1598,7 +1501,7 @@ void main() {
   });
 
   group('$OpenFoodAPIClient test ingredients', () {
-    const String barcode = _BARCODE_DANISH_BUTTER_COOKIES;
+    const String barcode = BARCODE_DANISH_BUTTER_COOKIES;
     // Ingredients for _BARCODE_DANISH_BUTTER_COOKIES
     const List<String> expectedIngredientLabels = <String>[
       'Buttergebäck',
@@ -1686,16 +1589,16 @@ void main() {
   test('get invalid barcodes', () async {
     final String url = InvalidBarcodes.getUrl();
     final http.Response response = await http.get(Uri.parse(url));
-    expect(response.statusCode, _HTTP_OK);
+    expect(response.statusCode, HTTP_OK);
     final String jsonString = response.body;
     InvalidBarcodes invalidBarcodes =
         InvalidBarcodes.loadFromJSONString(jsonString);
     assert(invalidBarcodes.isBlacklisted('15600703'));
-    assert(!invalidBarcodes.isBlacklisted(_BARCODE_DANISH_BUTTER_COOKIES));
+    assert(!invalidBarcodes.isBlacklisted(BARCODE_DANISH_BUTTER_COOKIES));
 
     invalidBarcodes = InvalidBarcodes.base();
     assert(invalidBarcodes.isBlacklisted('15600703'));
-    assert(!invalidBarcodes.isBlacklisted(_BARCODE_DANISH_BUTTER_COOKIES));
+    assert(!invalidBarcodes.isBlacklisted(BARCODE_DANISH_BUTTER_COOKIES));
   });
 
   test('get images freshness', () async {
@@ -1707,7 +1610,7 @@ void main() {
     ];
     final ProductResult productResult = await OpenFoodAPIClient.getProduct(
       ProductQueryConfiguration(
-        _BARCODE_DANISH_BUTTER_COOKIES,
+        BARCODE_DANISH_BUTTER_COOKIES,
         languages: languages,
         fields: [ProductField.IMAGES_FRESHNESS_IN_LANGUAGES],
       ),
@@ -1728,7 +1631,7 @@ void main() {
   });
 
   test('get product uri', () async {
-    const String barcode = _BARCODE_DANISH_BUTTER_COOKIES;
+    const String barcode = BARCODE_DANISH_BUTTER_COOKIES;
     expect(
       OpenFoodAPIClient.getProductUri(
         barcode,
@@ -1859,6 +1762,23 @@ void main() {
     expect(result.product?.packaging, 'de:In einer Plastikflasche');
     expect(result.product?.packagingTags, ['de:in-einer-plastikflasche']);
     expect(result.product?.quantity, '5.5 Liter');
+  });
+
+  test('get new product fields', () async {
+    final ProductQueryConfiguration configuration = ProductQueryConfiguration(
+      BARCODE_DANISH_BUTTER_COOKIES,
+      fields: [
+        ProductField.COMPARED_TO_CATEGORY,
+      ],
+    );
+
+    final ProductResult result = await OpenFoodAPIClient.getProduct(
+      configuration,
+    );
+
+    expect(result.status, 1);
+    expect(result.product, isNotNull);
+    expect(result.product!.comparedToCategory, isNotNull);
   });
 
   group('no nutrition data', () {