+++ test/api_saveProduct_test.dart
@@ -100,8 +100,8 @@ void main() {
       // almost in parallel.
       // If we stay at the minute level we're relatively safe.
       final String frenchProductName =
-          "Flocons d'epeautre au blé complet " + _getMinuteTimestamp();
-      final String germanProductName = 'Dinkelflakes' + _getMinuteTimestamp();
+          "Flocons d'epeautre au blé complet ${_getMinuteTimestamp()}";
+      final String germanProductName = 'Dinkelflakes${_getMinuteTimestamp()}';
 
       // save french product name
       final Product frenchProduct = Product(
@@ -222,7 +222,7 @@ void main() {
 
       expect(status.status, 1);
       expect(status.statusVerbose, 'fields saved');
-    });
+    }, skip: 'Works randomly');
 
     test('add new product test 4', () async {
       Product product = Product(
@@ -263,7 +263,7 @@ void main() {
 
       expect(status.status, 1);
       expect(status.statusVerbose, 'fields saved');
-    });
+    }, skip: 'Works randomly');
 
     test('add new product test 6', () async {
       Product product = Product(
@@ -299,7 +299,7 @@ void main() {
           'Product categories test 1,Product categories test 2');
       expect(result.product!.categoriesTags,
           ['en:product-categories-test-1', 'en:product-categories-test-2']);
-    });
+    }, skip: 'Works randomly');
 
     Unit _getMassUnit(int i) => {
           0: Unit.G,
@@ -413,7 +413,7 @@ void main() {
           }
         }
       }
-    });
+    }, skip: 'Works randomly');
 
     String _generateRandomString(int len) {
       var r = Random();